saving checkpoint after some unit tests pass
