+++ contracts/cw721-base-ibc/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cw721-ibc"
+name = "cw721-base-ibc"
 version = "0.13.2"
 authors = [
   "Ethan Frey <ethanfrey@users.noreply.github.com>",
@@ -31,7 +31,7 @@ library = []
 [dependencies]
 cw-utils = "0.13.2"
 cw2 = "0.13.2"
-cw721 = { path = "../../packages/cw721", version = "0.13.2" }
+cw721-ibc = { path = "../../packages/cw721-ibc", version = "0.13.2" }
 cw-storage-plus = "0.13.2"
 cosmwasm-std = { version = "1.0.0-beta8" }
 schemars = "0.8"
