+++ contracts/cw721-base-ibc/schema/approval_response.json
