+++ contracts/cw721-base-ibc/src/execute.rs
@@ -4,7 +4,7 @@ use serde::Serialize;
 use cosmwasm_std::{Binary, Deps, DepsMut, Env, MessageInfo, Response, StdResult};
 
 use cw2::set_contract_version;
-use cw721::{ContractInfoResponse, CustomMsg, Cw721Execute, Cw721ReceiveMsg, Expiration};
+use cw721_ibc::{ContractInfoResponse, CustomMsg, Cw721Execute, Cw721ReceiveMsg, Expiration};
 
 use crate::error::ContractError;
 use crate::msg::{ExecuteMsg, InstantiateMsg, MintMsg};
@@ -50,10 +50,11 @@ where
             ExecuteMsg::Approve {
                 spender,
                 token_id,
+                class_id, 
                 expires,
-            } => self.approve(deps, env, info, spender, token_id, expires),
-            ExecuteMsg::Revoke { spender, token_id } => {
-                self.revoke(deps, env, info, spender, token_id)
+            } => self.approve(deps, env, info, spender, token_id, class_id, expires),
+            ExecuteMsg::Revoke { spender, token_id, class_id } => {
+                self.revoke(deps, env, info, spender, token_id, class_id)
             }
             ExecuteMsg::ApproveAll { operator, expires } => {
                 self.approve_all(deps, env, info, operator, expires)
@@ -62,13 +63,15 @@ where
             ExecuteMsg::TransferNft {
                 recipient,
                 token_id,
-            } => self.transfer_nft(deps, env, info, recipient, token_id),
+                class_id
+            } => self.transfer_nft(deps, env, info, recipient, token_id, class_id),
             ExecuteMsg::SendNft {
                 contract,
                 token_id,
+                class_id,
                 msg,
-            } => self.send_nft(deps, env, info, contract, token_id, msg),
-            ExecuteMsg::Burn { token_id } => self.burn(deps, env, info, token_id),
+            } => self.send_nft(deps, env, info, contract, token_id, class_id, msg),
+            ExecuteMsg::Burn { token_id, class_id } => self.burn(deps, env, info, token_id, class_id),
         }
     }
 }
@@ -100,7 +103,7 @@ where
             extension: msg.extension,
         };
         self.tokens
-            .update(deps.storage, &msg.token_id, |old| match old {
+            .update(deps.storage, (&msg.token_id, &msg.class_id), |old| match old {
                 Some(_) => Err(ContractError::Claimed {}),
                 None => Ok(token),
             })?;
@@ -111,7 +114,9 @@ where
             .add_attribute("action", "mint")
             .add_attribute("minter", info.sender)
             .add_attribute("owner", msg.owner)
-            .add_attribute("token_id", msg.token_id))
+            .add_attribute("token_id", msg.token_id)
+            .add_attribute("class_id", msg.class_id)
+        )
     }
 }
 
@@ -129,14 +134,17 @@ where
         info: MessageInfo,
         recipient: String,
         token_id: String,
+        class_id: String
     ) -> Result<Response<C>, ContractError> {
-        self._transfer_nft(deps, &env, &info, &recipient, &token_id)?;
+        self._transfer_nft(deps, &env, &info, &recipient, &token_id, &class_id)?;
 
         Ok(Response::new()
             .add_attribute("action", "transfer_nft")
             .add_attribute("sender", info.sender)
             .add_attribute("recipient", recipient)
-            .add_attribute("token_id", token_id))
+            .add_attribute("token_id", token_id)
+            .add_attribute("class_id", class_id)
+        )
     }
 
     fn send_nft(
@@ -146,10 +154,11 @@ where
         info: MessageInfo,
         contract: String,
         token_id: String,
+        class_id: String,
         msg: Binary,
     ) -> Result<Response<C>, ContractError> {
         // Transfer token
-        self._transfer_nft(deps, &env, &info, &contract, &token_id)?;
+        self._transfer_nft(deps, &env, &info, &contract, &token_id, &class_id)?;
 
         let send = Cw721ReceiveMsg {
             sender: info.sender.to_string(),
@@ -163,7 +172,9 @@ where
             .add_attribute("action", "send_nft")
             .add_attribute("sender", info.sender)
             .add_attribute("recipient", contract)
-            .add_attribute("token_id", token_id))
+            .add_attribute("token_id", token_id)
+            .add_attribute("class_id", class_id)
+        )
     }
 
     fn approve(
@@ -173,15 +184,18 @@ where
         info: MessageInfo,
         spender: String,
         token_id: String,
+        class_id: String,
         expires: Option<Expiration>,
     ) -> Result<Response<C>, ContractError> {
-        self._update_approvals(deps, &env, &info, &spender, &token_id, true, expires)?;
+        self._update_approvals(deps, &env, &info, &spender, &token_id, &class_id, true, expires)?;
 
         Ok(Response::new()
             .add_attribute("action", "approve")
             .add_attribute("sender", info.sender)
             .add_attribute("spender", spender)
-            .add_attribute("token_id", token_id))
+            .add_attribute("token_id", token_id)
+            .add_attribute("class_id", class_id)
+        )
     }
 
     fn revoke(
@@ -191,14 +205,17 @@ where
         info: MessageInfo,
         spender: String,
         token_id: String,
+        class_id: String,
     ) -> Result<Response<C>, ContractError> {
-        self._update_approvals(deps, &env, &info, &spender, &token_id, false, None)?;
+        self._update_approvals(deps, &env, &info, &spender, &token_id, &class_id, false, None)?;
 
         Ok(Response::new()
             .add_attribute("action", "revoke")
             .add_attribute("sender", info.sender)
             .add_attribute("spender", spender)
-            .add_attribute("token_id", token_id))
+            .add_attribute("token_id", token_id)
+            .add_attribute("class_id", class_id)
+        )
     }
 
     fn approve_all(
@@ -249,17 +266,20 @@ where
         env: Env,
         info: MessageInfo,
         token_id: String,
+        class_id: String
     ) -> Result<Response<C>, ContractError> {
-        let token = self.tokens.load(deps.storage, &token_id)?;
+        let token = self.tokens.load(deps.storage, (&token_id, &class_id))?;
         self.check_can_send(deps.as_ref(), &env, &info, &token)?;
 
-        self.tokens.remove(deps.storage, &token_id)?;
+        self.tokens.remove(deps.storage, (&token_id, &class_id))?;
         self.decrement_tokens(deps.storage)?;
 
         Ok(Response::new()
             .add_attribute("action", "burn")
             .add_attribute("sender", info.sender)
-            .add_attribute("token_id", token_id))
+            .add_attribute("token_id", token_id)
+            .add_attribute("class_id", class_id)
+        )
     }
 }
 
@@ -276,14 +296,15 @@ where
         info: &MessageInfo,
         recipient: &str,
         token_id: &str,
+        class_id: &str,
     ) -> Result<TokenInfo<T>, ContractError> {
-        let mut token = self.tokens.load(deps.storage, token_id)?;
+        let mut token = self.tokens.load(deps.storage, (&token_id, &class_id))?;
         // ensure we have permissions
         self.check_can_send(deps.as_ref(), env, info, &token)?;
         // set owner and remove existing approvals
         token.owner = deps.api.addr_validate(recipient)?;
         token.approvals = vec![];
-        self.tokens.save(deps.storage, token_id, &token)?;
+        self.tokens.save(deps.storage, (token_id, &class_id), &token)?;
         Ok(token)
     }
 
@@ -295,11 +316,12 @@ where
         info: &MessageInfo,
         spender: &str,
         token_id: &str,
+        class_id: &str,
         // if add == false, remove. if add == true, remove then set with this expiration
         add: bool,
         expires: Option<Expiration>,
     ) -> Result<TokenInfo<T>, ContractError> {
-        let mut token = self.tokens.load(deps.storage, token_id)?;
+        let mut token = self.tokens.load(deps.storage, (&token_id, &class_id))?;
         // ensure we have permissions
         self.check_can_approve(deps.as_ref(), env, info, &token)?;
 
@@ -325,7 +347,7 @@ where
             token.approvals.push(approval);
         }
 
-        self.tokens.save(deps.storage, token_id, &token)?;
+        self.tokens.save(deps.storage, (&token_id, &class_id), &token)?;
 
         Ok(token)
     }
