+++ packages/cw721-ibc/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cw721"
+name = "cw721-ibc"
 version = "0.13.2"
 authors = [
     "Ethan Frey <ethanfrey@users.noreply.github.com>",
