+++ packages/cw721-ibc/schema/cw721_query_msg.json
