+++ packages/cw721-ibc/schema/nft_info_response.json
